update erpc version

+++ checker.go
@@ -9,12 +9,12 @@ import (
 	pb "github.com/es-hs/erpc/authz"
 )
 
-func CheckPermission(userId uint, shopId uint, role string) (bool, error) {
+func CheckPermission(userId uint64, shopId uint64, role string) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.CheckPermission(ctx, &pb.CheckPermissionRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 		Act:    role,
 	})
 	if err != nil {
@@ -23,12 +23,12 @@ func CheckPermission(userId uint, shopId uint, role string) (bool, error) {
 	return r.Result, err
 }
 
-func AddRoleToDomain(userId uint, shopId uint, role string) (bool, error) {
+func AddRoleToDomain(userId uint64, shopId uint64, role string) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.AddRoleToDomain(ctx, &pb.AddRoleToDomainRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 		Act:    role,
 	})
 	if err != nil {
@@ -37,12 +37,12 @@ func AddRoleToDomain(userId uint, shopId uint, role string) (bool, error) {
 	return r.Result, err
 }
 
-func GetRolesInDomain(userId uint, shopId uint) ([]string, error) {
+func GetRolesInDomain(userId uint64, shopId uint64) ([]string, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.GetRolesInDomain(ctx, &pb.GetRolesInDomainRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 	})
 	if err != nil {
 		return nil, err
@@ -50,12 +50,12 @@ func GetRolesInDomain(userId uint, shopId uint) ([]string, error) {
 	return r.Roles, err
 }
 
-func GetImplicitRolesInDomain(userId uint, shopId uint) ([]string, error) {
+func GetImplicitRolesInDomain(userId uint64, shopId uint64) ([]string, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.GetImplicitRolesInDomain(ctx, &pb.GetImplicitRolesInDomainRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 	})
 	if err != nil {
 		return nil, err
@@ -63,25 +63,25 @@ func GetImplicitRolesInDomain(userId uint, shopId uint) ([]string, error) {
 	return r.Roles, err
 }
 
-func GenerateOwnerRole(userId uint, shopId uint) (uint, error) {
+func GenerateOwnerRole(userId uint64, shopId uint64) (uint64, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.GenerateOwnerRole(ctx, &pb.GenerateOwnerRoleRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 	})
 	if err != nil {
 		return 1, err
 	}
-	return uint(r.Code), nil
+	return uint64(r.Code), nil
 }
 
-func AddRolesForUserToDomain(userId uint, shopId uint, roles []string) (bool, error) {
+func AddRolesForUserToDomain(userId uint64, shopId uint64, roles []string) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.AddRolesForUserToDomain(ctx, &pb.AddRolesForUserToDomainRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 		Act:    roles,
 	})
 	if err != nil {
@@ -90,12 +90,12 @@ func AddRolesForUserToDomain(userId uint, shopId uint, roles []string) (bool, er
 	return r.Result, nil
 }
 
-func RemoveRolesFromDomain(userId uint, shopId uint, roles []string) (bool, error) {
+func RemoveRolesFromDomain(userId uint64, shopId uint64, roles []string) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.RemoveRolesFromDomain(ctx, &pb.RemoveRolesFromDomainRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 		Act:    roles,
 	})
 	if err != nil {
@@ -104,12 +104,12 @@ func RemoveRolesFromDomain(userId uint, shopId uint, roles []string) (bool, erro
 	return r.Result, nil
 }
 
-func RemoveRoleFromDomain(userId uint, shopId uint, role string) (bool, error) {
+func RemoveRoleFromDomain(userId uint64, shopId uint64, role string) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.RemoveRoleFromDomain(ctx, &pb.RemoveRoleFromDomainRequest{
-		UserId: int64(userId),
-		ShopId: int64(shopId),
+		UserId: userId,
+		ShopId: shopId,
 		Act:    role,
 	})
 	if err != nil {
@@ -118,13 +118,13 @@ func RemoveRoleFromDomain(userId uint, shopId uint, role string) (bool, error) {
 	return r.Result, err
 }
 
-func GetAllUsersByDomain(shopId uint) ([]uint, []uint, error) {
-	var userIds []uint
-	var partnerIds []uint
+func GetAllUsersByDomain(shopId uint64) ([]uint64, []uint64, error) {
+	var userIds []uint64
+	var partnerIds []uint64
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.GetAllUsersByDomain(ctx, &pb.GetAllUsersByDomainRequest{
-		ShopId: int64(shopId),
+		ShopId: shopId,
 	})
 	if err != nil {
 		return nil, nil, err
@@ -132,23 +132,23 @@ func GetAllUsersByDomain(shopId uint) ([]uint, []uint, error) {
 	for k := range r.UserIds {
 		if strings.Contains(r.UserIds[k], "user_") {
 			uID, _ := strconv.Atoi(r.UserIds[k][5:])
-			userIds = append(userIds, uint(uID))
+			userIds = append(userIds, uint64(uID))
 		}
 		if strings.Contains(r.UserIds[k], "partner_") {
 			uID, _ := strconv.Atoi(r.UserIds[k][8:])
-			partnerIds = append(partnerIds, uint(uID))
+			partnerIds = append(partnerIds, uint64(uID))
 		}
 	}
 	return userIds, partnerIds, err
 }
 
-func GetUsersForRoleInDomain(shopId uint, role string) ([]uint, []uint, error) {
-	var userIds []uint
-	var partnerIds []uint
+func GetUsersForRoleInDomain(shopId uint64, role string) ([]uint64, []uint64, error) {
+	var userIds []uint64
+	var partnerIds []uint64
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.GetUsersForRoleInDomain(ctx, &pb.GetUsersForRoleInDomainRequest{
-		ShopId: int64(shopId),
+		ShopId: shopId,
 		Role:   role,
 	})
 	if err != nil {
@@ -157,22 +157,22 @@ func GetUsersForRoleInDomain(shopId uint, role string) ([]uint, []uint, error) {
 	for k := range r.UserIds {
 		if strings.Contains(r.UserIds[k], "user_") {
 			uID, _ := strconv.Atoi(r.UserIds[k][5:])
-			userIds = append(userIds, uint(uID))
+			userIds = append(userIds, uint64(uID))
 		}
 		if strings.Contains(r.UserIds[k], "partner_") {
 			uID, _ := strconv.Atoi(r.UserIds[k][8:])
-			partnerIds = append(partnerIds, uint(uID))
+			partnerIds = append(partnerIds, uint64(uID))
 		}
 	}
 
 	return userIds, partnerIds, err
 }
 
-func DeleteDomains(shopId uint) (bool, error) {
+func DeleteDomains(shopId uint64) (bool, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
 	r, err := C.DeleteDomains(ctx, &pb.DeleteDomainsRequest{
-		ShopIds: []int64{int64(shopId)},
+		ShopIds: []uint64{shopId},
 	})
 	if err != nil {
 		return false, err
